+++ scripts/curate_bibliography/generate_packages_bib.R
@@ -44,6 +44,7 @@ SOFTWARE_TOOLS <- c(
   "igraph",
   "RSpectra",
   "numDeriv",
+  "mvtnorm",
   "compositions",
   "marqLevAlg",
   "ComplexHeatmap",